Refactor Login Page to point to new API endpoint. Create Custom Hook to manage User Creation
